+ Optimize a few div.c + Remove some warnings in asin.c, atan.c. + Include limits.h before gmp-impl.h in mpfr-impl.h + mpfr-tests.h doesn't include anumore standard include files. + Test files include only standard includes and mpfr-tests.h. + Add some forgotten mpfr_clear in mpf[r]_compat. + Remove some warnings in the tests.
[From SVN r2576 (trunk)]

+++ tests/tget_str.c
@@ -22,10 +22,7 @@ MA 02111-1307, USA. */
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "gmp.h"
-#include "gmp-impl.h"
-#include "mpfr.h"
-#include "mpfr-impl.h"
+
 #include "mpfr-test.h"
 
 static void
@@ -400,7 +397,7 @@ check_special (int b, mp_prec_t p)
     {
       /* x = b^i (exact) */
       for (r=0; r<4; r++)
-        for (m=(i < 3) ? 2 : i-1; m<=i+1; m++)
+        for (m= (i<3)? 2 : i-1 ; (int) m <= i+1 ; m++)
           {
             mpfr_get_str (s, &e, b, m, x, r);
             /* s should be 1 followed by (m-1) zeros, and e should be i+1 */
